+++ src/renderer/stores/settings-store.ts
@@ -1,13 +1,18 @@
 import create from "zustand";
 
-import { SettingsMeta } from "common/types";
+import { Settings, SettingsMeta } from "common/types";
 import { defaultSettings } from "common/settings";
 import { getSettings as invokeGetSettings, saveSettings } from "../ipc";
+import { MacroActionEffect, getMacroActionValue, MacroAction } from "../command-center/macros";
+import { getLevaConfigField } from "../command-center/global-settings";
 
+import lSet from "lodash.set";
 
 export type SettingsStore = SettingsMeta & {
-  save: (data: any) => Promise<void>;
+  save: (data: Partial<Settings>) => Promise<void>;
+  set: (data: Partial<Settings>) => Promise<void>;
   load: () => Promise<SettingsMeta>;
+  doMacroAction: (action: MacroAction) => void;
 };
 
 export const useSettingsStore = create<SettingsStore>((set, get) => ({
@@ -17,6 +22,9 @@ export const useSettingsStore = create<SettingsStore>((set, get) => ({
   errors: [],
   enabledPlugins: [],
   disabledPlugins: [],
+  set: async (settings) => {
+    set((state) => ({ data: { ...state.data, ...settings } }));
+  },
   save: async (settings) => {
     set((state) => ({ data: { ...state.data, ...settings } }));
     await saveSettings(get().data);
@@ -26,6 +34,21 @@ export const useSettingsStore = create<SettingsStore>((set, get) => ({
     set(settings);
     return settings;
   },
+  doMacroAction: async (action) => {
+    if (action.effect === MacroActionEffect.CallMethod) {
+      return;
+    }
+
+    const field = getLevaConfigField(get().data, action.field!) as any;
+    if (field === undefined) {
+      return;
+    }
+
+    const value = getMacroActionValue(action, field.value, field.step, field.min, field.max);
+    const newSettings = { ...get().data };
+    lSet(newSettings, action.field!, value);
+    get().set(newSettings);
+  }
 }));