+++ src/renderer/utils/ses-util.ts
@@ -1,0 +1,24 @@
+import * as THREE from "three";
+import * as postprocessing from "postprocessing"
+import cameraControls from "camera-controls"
+import * as enums from "common/enums";
+import Janitor from "@utils/janitor";
+import { Layers } from "../render/layers";
+import { CSS2DObject } from "three/examples/jsm/renderers/CSS2DRenderer";
+const STDLIB = {
+    CSS2DObject
+}
+
+export const createCompartment = (env: {} = {}) => {
+
+    const modules = { THREE, STDLIB, postprocessing, Janitor, Layers, enums, cameraControls }
+
+    return new Compartment({
+        ...env,
+        console: harden(console),
+        Math,
+        ...modules,
+        Image: harden(Image),
+    });
+
+}